rename imports in service/controller/aws/v7patch1: github.com/giantswarm/cluster-operator/service/controller/aws/v7 > github.com/giantswarm/cluster-operator/service/controller/aws/v7patch1
