+++ service/controller/aws/v7patch1/resource/chartconfig/update.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/giantswarm/cluster-operator/pkg/v7/chartconfig"
 	"github.com/giantswarm/cluster-operator/pkg/v7patch1/key"
-	awskey "github.com/giantswarm/cluster-operator/service/controller/aws/v7/key"
+	awskey "github.com/giantswarm/cluster-operator/service/controller/aws/v7patch1/key"
 )
 
 func (r *Resource) ApplyUpdateChange(ctx context.Context, obj, updateChange interface{}) error {
