+++ service/controller/aws/v7patch1/resource_set.go
@@ -26,10 +26,10 @@ import (
 	"github.com/giantswarm/cluster-operator/pkg/v7/resource/clustercr"
 	"github.com/giantswarm/cluster-operator/pkg/v7/resource/encryptionkey"
 	"github.com/giantswarm/cluster-operator/pkg/v7/resource/namespace"
-	"github.com/giantswarm/cluster-operator/service/controller/aws/v7/key"
-	"github.com/giantswarm/cluster-operator/service/controller/aws/v7/resource/awsconfig"
-	"github.com/giantswarm/cluster-operator/service/controller/aws/v7/resource/chartconfig"
-	"github.com/giantswarm/cluster-operator/service/controller/aws/v7/resource/configmap"
+	"github.com/giantswarm/cluster-operator/service/controller/aws/v7patch1/key"
+	"github.com/giantswarm/cluster-operator/service/controller/aws/v7patch1/resource/awsconfig"
+	"github.com/giantswarm/cluster-operator/service/controller/aws/v7patch1/resource/chartconfig"
+	"github.com/giantswarm/cluster-operator/service/controller/aws/v7patch1/resource/configmap"
 )
 
 // ResourceSetConfig contains necessary dependencies and settings for
